move G to common
