+++ src/climatevision/generator/heat2030/energy_general/g.py
@@ -7,18 +7,6 @@ from ...utils import div
 
 
 @dataclass(kw_only=True)
-class G:
-    cost_wage: float
-    demand_emplo: float
-    demand_emplo_com: float
-    demand_emplo_new: float
-    invest: float
-    invest_com: float
-    invest_pa: float
-    invest_pa_com: float
-
-
-@dataclass(kw_only=True)
 class GStorage:
     cost_wage: float
     demand_emplo: float
