+++ src/climatevision/generator/transport2030/energy_general/g.py
@@ -66,15 +66,3 @@ class GPlanning:
             invest_pa_com=invest_pa_com,
             ratio_wage_to_emplo=ratio_wage_to_emplo,
         )
-
-
-@dataclass(kw_only=True)
-class G:
-    cost_wage: float
-    demand_emplo: float
-    demand_emplo_com: float
-    demand_emplo_new: float
-    invest: float
-    invest_com: float
-    invest_pa: float
-    invest_pa_com: float
